+++ include/skl.h
@@ -62,7 +62,7 @@
 #endif
 
 #if defined(__riscv_xsfvqdotq)
-#include "../src/gemm/xsfvqdotq/gemm_i8_i8pc_i32_xsfvqdotq.h"
+#include "../src/gemm/xsfvqdotq/gemm_i8rcp_i8pc_i32_xsfvqdotq.h"
 #endif
 
 /*
